Potential fix for code scanning alert no. 2: Implicit narrowing conversion in compound assignment

+++ src/main/java/com/example/miscfeatures/command/FindItem.java
@@ -241,7 +241,7 @@ public final class FindItem {
             }
         });
 
-        int grandTotal = 0;
+        long grandTotal = 0L;
         for (Map.Entry<ItemSearchTarget, List<ContainerMatch>> entry : results.entrySet()) {
             ItemSearchTarget target = entry.getKey();
             List<ContainerMatch> matches = entry.getValue();
